refactor: move PiloteDateFormatter
