+++ src/client/components/PageRapportsHebdomadaires/PageRapportsHebdomadaires.tsx
@@ -3,7 +3,7 @@ import { parseAsString, useQueryState } from "nuqs";
 import api from "@/server/infrastructure/api/trpc/api";
 import BarreLatérale from "@/components/_commons/BarreLatérale/BarreLatérale";
 import BarreLatéraleEncart from "@/components/_commons/BarreLatérale/BarreLatéraleEncart/BarreLatéraleEncart";
-import { PiloteDateFormatter } from "@/server/rapports-hebdomadaires/infrastructure/adapters/PiloteDateFormatter";
+import { PiloteDateFormatter } from "@/utils/PiloteDateFormatter";
 import { Bouton } from "@/components/_commons/Bouton/Bouton";
 import { RapportDetail } from "./RapportDetail";
