+++ src/client/components/PageRapportsHebdomadaires/RapportDetail.tsx
@@ -1,6 +1,6 @@
 import api from "@/server/infrastructure/api/trpc/api";
 import Bloc from "@/components/_commons/Bloc/Bloc";
-import { PiloteDateFormatter } from "@/server/rapports-hebdomadaires/infrastructure/adapters/PiloteDateFormatter";
+import { PiloteDateFormatter } from "@/utils/PiloteDateFormatter";
 import { TitreSection } from "./TitreSection";
 import { BlocEtatVide } from "./BlocEtatVide";
 import { TableauUtilisateurs } from "./TableauUtilisateurs";
