+++ src/client/components/_commons/Widget/WidgetCartographieMeteo/RepartitionNiveauxDeConfiance.tsx
@@ -80,8 +80,7 @@ export const RepartitionNiveauxDeConfiance = ({
                     couleur={couleur}
                     onSupprimer={
                       !estInitial
-                        ? () =>
-                            onSupprimerTerritoire(territoire.territoireCode)
+                        ? () => onSupprimerTerritoire(territoire.territoireCode)
                         : undefined
                     }
                   />
